fixes h5 writer
